# backend/uzum_openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Uzum market seller openapi",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api-seller.uzum.uz/api/seller-openapi/",
      "description": "OpenAPI для продавцов UZUM MARKET"
    }
  ],
  "security": [
    {
      "TokenAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/finance/orders": {
      "get": {
        "summary": "Получение списка заказов.",
        "description": "Этот метод позволяет получить список продаж.",
        "operationId": "getFinanceOrders",
        "tags": [
          "Finance"
        ],
        "parameters": [
          {
            "name": "page",
            "description": "Номер страницы для постраничного просмотра накладных поставки.",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "description": "Количество записей на странице (постраничный просмотр).",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "group",
            "description": "Группировать заказы или нет. В зависимости от состояния этого флага, будет возвращаться то или иное наполнения поля orderItems. >\nЕсли флаг true: вернется `ProductGroupedSellerItem`\nЕсли флаг false: вернется `SellerOrderItemDto`",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "dateFrom",
            "description": "Дата, с которой необходимо показывать заказы",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "dateTo",
            "description": "Дата, с которой необходимо показывать заказы",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "statuses",
            "description": "Статусы заказов (к выводу средств (TO_WITHDRAW), в обработке (PROCESSING), отменен (CANCELED), частично отменен( PARTIALLY_CANCELLED))",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "TO_WITHDRAW",
                  "PROCESSING",
                  "CANCELED",
                  "PARTIALLY_CANCELLED"
                ]
              }
            }
          },
          {
            "name": "shopIds",
            "description": "Список магазинов, для которых необходимо показывать заказы",
            "in": "query",
            "required": true,
            "schema": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "x-ratelimit-remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "x-ratelimit-replenish-rate": {
                "$ref": "#/components/headers/X-RateLimit-Replenish-Rate"
              },
              "x-ratelimit-burst-capacity": {
                "$ref": "#/components/headers/X-RateLimit-Burst-Capacity"
              },
              "x-ratelimit-requested-tokens": {
                "$ref": "#/components/headers/X-RateLimit-Requested-Tokens"
              },
              "x-ratelimit-limit-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Limit-Per-Day"
              },
              "x-ratelimit-remaining-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Remaining-Per-Day"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceOrderItemsDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "headers": {
              "x-ratelimit-remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "x-ratelimit-replenish-rate": {
                "$ref": "#/components/headers/X-RateLimit-Replenish-Rate"
              },
              "x-ratelimit-burst-capacity": {
                "$ref": "#/components/headers/X-RateLimit-Burst-Capacity"
              },
              "x-ratelimit-requested-tokens": {
                "$ref": "#/components/headers/X-RateLimit-Requested-Tokens"
              },
              "x-ratelimit-limit-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Limit-Per-Day"
              },
              "x-ratelimit-remaining-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Remaining-Per-Day"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseObject"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fbs/order/{orderId}": {
      "get": {
        "summary": "Получение информации о заказе",
        "description": "Возвращает информацию о заказе продавца по идентификатору.",
        "operationId": "getFbsOrderByOrderId",
        "tags": [
          "FBS"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Идентификатор заказа продавца."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "x-ratelimit-remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "x-ratelimit-replenish-rate": {
                "$ref": "#/components/headers/X-RateLimit-Replenish-Rate"
              },
              "x-ratelimit-burst-capacity": {
                "$ref": "#/components/headers/X-RateLimit-Burst-Capacity"
              },
              "x-ratelimit-requested-tokens": {
                "$ref": "#/components/headers/X-RateLimit-Requested-Tokens"
              },
              "x-ratelimit-limit-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Limit-Per-Day"
              },
              "x-ratelimit-remaining-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Remaining-Per-Day"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseSellerOrderDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {
              "x-ratelimit-remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "x-ratelimit-replenish-rate": {
                "$ref": "#/components/headers/X-RateLimit-Replenish-Rate"
              },
              "x-ratelimit-burst-capacity": {
                "$ref": "#/components/headers/X-RateLimit-Burst-Capacity"
              },
              "x-ratelimit-requested-tokens": {
                "$ref": "#/components/headers/X-RateLimit-Requested-Tokens"
              },
              "x-ratelimit-limit-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Limit-Per-Day"
              },
              "x-ratelimit-remaining-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Remaining-Per-Day"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/fbs/orders": {
      "get": {
        "summary": "Получение заказов продавца",
        "description": "Возвращает заказы продавца по статусу.",
        "operationId": "getFbsOrdersV2",
        "tags": [
          "FBS"
        ],
        "parameters": [
          {
            "name": "shopIds",
            "description": "Массив идентификаторов магазинов, для которых нужно получить заказы.",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          {
            "name": "status",
            "description": "Статус заказов, которые нужно получить.",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "CREATED",
              "enum": [
                "CREATED",
                "PACKING",
                "PENDING_DELIVERY",
                "DELIVERING",
                "DELIVERED",
                "ACCEPTED_AT_DP",
                "DELIVERED_TO_CUSTOMER_DELIVERY_POINT",
                "COMPLETED",
                "CANCELED",
                "PENDING_CANCELLATION",
                "RETURNED"
              ]
            }
          },
          {
            "name": "scheme",
            "description": "Тип заказа. Если ничего не передать, вернутся и FBS и DBS",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "FBS",
                "DBS"
              ]
            }
          },
          {
            "name": "dateFrom",
            "description": "Дата, с которой необходимо показывать заказы",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "dateTo",
            "description": "Дата, по которую необходимо показывать заказы",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "description": "Номер страницы для получения данных (начинается с 0)",
            "in": "query",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "description": "Количество возвратов на странице (максимум 50).",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "x-ratelimit-remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "x-ratelimit-replenish-rate": {
                "$ref": "#/components/headers/X-RateLimit-Replenish-Rate"
              },
              "x-ratelimit-burst-capacity": {
                "$ref": "#/components/headers/X-RateLimit-Burst-Capacity"
              },
              "x-ratelimit-requested-tokens": {
                "$ref": "#/components/headers/X-RateLimit-Requested-Tokens"
              },
              "x-ratelimit-limit-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Limit-Per-Day"
              },
              "x-ratelimit-remaining-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Remaining-Per-Day"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseSellerOrdersDto"
                }
              }
            }
          },
          "400": {
            "description": "seller-order-12 - dateTo is before dateFrom",
            "headers": {
              "x-ratelimit-remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "x-ratelimit-replenish-rate": {
                "$ref": "#/components/headers/X-RateLimit-Replenish-Rate"
              },
              "x-ratelimit-burst-capacity": {
                "$ref": "#/components/headers/X-RateLimit-Burst-Capacity"
              },
              "x-ratelimit-requested-tokens": {
                "$ref": "#/components/headers/X-RateLimit-Requested-Tokens"
              },
              "x-ratelimit-limit-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Limit-Per-Day"
              },
              "x-ratelimit-remaining-per-day": {
                "$ref": "#/components/headers/X-RateLimit-Remaining-Per-Day"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponseSellerOrdersDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/fbs/invoice/{invoiceId}/orders": {
        "get": {
            "summary": "Получить FBS заказы по идентификатору накладной",
            "description": "Получить FBS заказы по идентификатору накладной",
            "operationId": "getFbsOrdersByInvoiceId",
            "tags": [
                "FBS Invoice"
            ],
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "Идентификатор накладной",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                {
                    "name": "Accept-Language",
                    "in": "header",
                    "description": "Язык локализации. Поддерживаемые языки: ru, uz. По умолчанию: uz",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": ["ru", "uz"],
                        "default": "uz"
                    }
                }
            ],
            "responses": {
                "403": {
                    "description": "fbs-2-seller-access-denied - У продавца нет доступа",
                    "headers": {
                        "x-ratelimit-remaining": {
                            "$ref": "#/components/headers/X-RateLimit-Remaining"
                        },
                        "x-ratelimit-replenish-rate": {
                            "$ref": "#/components/headers/X-RateLimit-Replenish-Rate"
                        },
                        "x-ratelimit-burst-capacity": {
                            "$ref": "#/components/headers/X-RateLimit-Burst-Capacity"
                        },
                        "x-ratelimit-requested-tokens": {
                            "$ref": "#/components/headers/X-RateLimit-Requested-Tokens"
                        },
                        "x-ratelimit-limit-per-day": {
                            "$ref": "#/components/headers/X-RateLimit-Limit-Per-Day"
                        },
                        "x-ratelimit-remaining-per-day": {
                            "$ref": "#/components/headers/X-RateLimit-Remaining-Per-Day"
                        }
                    },
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GenericResponse"
                            }
                        }
                    }
                },
                "200": {
                    "description": "Success",
                    "headers": {
                        "x-ratelimit-remaining": {
                            "$ref": "#/components/headers/X-RateLimit-Remaining"
                        },
                        "x-ratelimit-replenish-rate": {
                            "$ref": "#/components/headers/X-RateLimit-Replenish-Rate"
                        },
                        "x-ratelimit-burst-capacity": {
                            "$ref": "#/components/headers/X-RateLimit-Burst-Capacity"
                        },
                        "x-ratelimit-requested-tokens": {
                            "$ref": "#/components/headers/X-RateLimit-Requested-Tokens"
                        },
                        "x-ratelimit-limit-per-day": {
                            "$ref": "#/components/headers/X-RateLimit-Limit-Per-Day"
                        },
                        "x-ratelimit-remaining-per-day": {
                            "$ref": "#/components/headers/X-RateLimit-Remaining-Per-Day"
                        }
                    },
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GenericResponseListFbsOrdersWithItemsDto"
                            }
                        }
                    }
                }
            }
        }
    },
...
  },
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Токен авторизации без префикса Bearer"
      }
    },
    "headers": {
      "X-RateLimit-Remaining": {
        "description": "Оставшееся количество запросов на ближайшую секунду",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      },
      "X-RateLimit-Replenish-Rate": {
        "description": "Скорость пополнения лимита запросов",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      },
      "X-RateLimit-Burst-Capacity": {
        "description": "Максимальное количество запросов, которые могут быть выполнены за одну секунду",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      },
      "X-RateLimit-Requested-Tokens": {
        "description": "Количество токенов, запрашиваемых за текущий запрос",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      },
      "X-RateLimit-Limit-Per-Day": {
        "description": "Лимит запросов в день",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      },
      "X-RateLimit-Remaining-Per-Day": {
        "description": "Оставшееся количество запросов в день",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "schemas": {
...
      "FbsOrdersWithItemsDto": {
            "required": [
                "fullPrice",
                "items",
                "orderId"
            ],
            "type": "object",
            "properties": {
                "orderId": {
                    "type": "integer",
                    "description": "Накладная",
                    "format": "int64"
                },
                "fullPrice": {
                    "type": "integer",
                    "description": "Сумма заказа",
                    "format": "int64"
                },
                "items": {
                    "type": "array",
                    "description": "Список элементов заказов",
                    "items": {
                        "$ref": "#/components/schemas/FbsInvoiceOrderItemDto"
                    }
                }
            },
            "description": "Response payload"
        },
      "GenericResponseListFbsOrdersWithItemsDto": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "array",
                    "description": "Response payload",
                    "items": {
                        "$ref": "#/components/schemas/FbsOrdersWithItemsDto"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                },
                "timestamp": {
                    "type": "string",
                    "format": "date-time"
                },
                "trace": {
                    "type": "string"
                },
                "error": {
                    "type": "string",
                    "readOnly": true,
                    "example": "null"
                }
            }
        },
...
      "GenericResponseObject": {
        "type": "object",
        "properties": {
          "payload": {
            "type": "object",
            "description": "Данные ответа, содержит информацию об объекте, возвращаемом сервисом."
          },
          "errors": {
            "type": "array",
            "description": "Список ошибок, если запрос не был успешно выполнен.",
            "items": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Время формирования ответа, формат даты и времени."
          },
          "trace": {
            "type": "string",
            "description": "Трассировка для отладки ошибок."
          },
          "error": {
            "type": "string",
            "description": "Сообщение об ошибке, если оно имеется.",
            "readOnly": true,
            "example": "null"
          }
        }
      },
...
      "ApiErrorResponse": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Код ошибки, который может помочь в определении причины."
          },
          "message": {
            "type": "string",
            "description": "Сообщение об ошибке, предназначенное для отображения пользователю."
          },
          "detailMessage": {
            "type": "string",
            "deprecated": true,
            "description": "Более подробное сообщение об ошибке. Возможно, устарело и больше не используется."
          },
          "payload": {
            "type": "object",
            "description": "Дополнительные данные, связанные с ошибкой."
          }
        }
      },
...
      "FinanceOrderItemsDto": {
        "type": "object",
        "properties": {
          "orderItems": {
            "type": "array",
            "description": "Список элементов заказа для финансовых расчетов",
            "items": {
              "$ref": "#/components/schemas/FinanceItemEntity"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int32",
            "description": "Общее количество элементов"
          }
        }
      },
      "FinanceItemEntity": {
        "type": "object",
        "description": "Базовый объект финансового элемента"
      },
...
      "StockDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Уникальный идентификатор склада"
          },
          "externalId": {
            "type": "string",
            "description": "Внешний идентификатор склада"
          },
          "title": {
            "type": "string",
            "description": "Название склада"
          },
          "address": {
            "type": "string",
            "description": "Адрес склада"
          },
          "timeFrom": {
            "type": "string",
            "description": "Время начала работы склада"
          },
          "timeTo": {
            "type": "string",
            "description": "Время окончания работы склада"
          },
          "poolSource": {
            "type": "string",
            "description": "Источник складских запасов"
          },
          "dimensionalGroups": {
            "type": "array",
            "description": "Список размерных групп, поддерживаемых складом",
            "items": {
              "$ref": "#/components/schemas/DimensionalGroupDto"
            }
          }
        }
      },
...
      "TimeSlotDto": {
        "type": "object",
        "properties": {
          "timeFrom": {
            "type": "string",
            "description": "Время начала временного интервала",
            "format": "date-time"
          },
          "timeTo": {
            "type": "string",
            "description": "Время окончания временного интервала",
            "format": "date-time"
          }
        },
        "description": "Модель временного интервала"
      },
...
      "GenericResponseSellerOrdersDto": {
        "type": "object",
        "description": "Общий ответ, содержащий информацию о заказах продавца.",
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/SellerOrdersDto"
          },
          "errors": {
            "type": "array",
            "description": "Список ошибок, возникших при обработке запроса.",
            "items": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Метка времени ответа."
          },
          "trace": {
            "type": "string",
            "description": "Трассировочная информация для отладки."
          },
          "error": {
            "type": "string",
            "readOnly": true,
            "example": "null",
            "description": "Описание ошибки, если она произошла."
          }
        }
      },
      "SellerOrdersDto": {
        "type": "object",
        "description": "Информация о заказах продавца.",
        "properties": {
          "orders": {
            "type": "array",
            "description": "Список заказов.",
            "items": {
              "$ref": "#/components/schemas/SellerOrderDto"
            }
          },
          "totalAmount": {
            "type": "integer",
            "description": "Общее количество заказов.",
            "format": "int64"
          }
        }
      },
      "SellerOrderDto": {
        "type": "object",
        "description": "Подробная информация о заказе продавца.",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Уникальный идентификатор заказа.",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "description": "Статус заказа.",
            "enum": [
              "CREATED",
              "PACKING",
              "DELIVERED",
              "DELIVERED_TO_CUSTOMER_DELIVERY_POINT",
              "COMPLETED",
              "CANCELED",
              "PENDING_CANCELLATION",
              "RETURNED"
            ]
          },
          "dateCreated": {
            "type": "string",
            "description": "Дата создания заказа.",
            "format": "date-time"
          },
          "acceptUntil": {
            "type": "string",
            "description": "Дата, до которой заказ может быть принят.",
            "format": "date-time"
          },
          "deliverUntil": {
            "type": "string",
            "description": "Дата, до которой заказ должен быть доставлен.",
            "format": "date-time"
          },
          "deliveringDate": {
            "type": "string",
            "description": "Дата перехода заказа в статус DELIVERING",
            "format": "date-time"
          },
          "deliveryDate": {
            "type": "string",
            "description": "Дата фактической доставки заказа.",
            "format": "date-time"
          },
          "acceptedDate": {
            "type": "string",
            "description": "Дата принятия заказа продавцом.",
            "format": "date-time"
          },
          "deliveredToDeliveryPointDate": {
            "type": "string",
            "description": "Дата доставки заказа в пункт доставки.",
            "format": "date-time"
          },
          "completedDate": {
            "type": "string",
            "description": "Дата завершения заказа.",
            "format": "date-time"
          },
          "dateCancelled": {
            "type": "string",
            "description": "Дата отмены заказа.",
            "format": "date-time"
          },
          "returnDate": {
            "type": "string",
            "description": "Дата возврата заказа.",
            "format": "date-time"
          },
          "cancelReason": {
            "type": "string",
            "description": "Причина отмены заказа."
          },
          "identifierRequired": {
            "type": "boolean",
            "description": "Требуется ли идентификатор для обработки заказа."
          },
          "price": {
            "type": "integer",
            "description": "Стоимость заказа.",
            "format": "int64"
          },
          "shopId": {
            "type": "integer",
            "description": "Идентификатор магазина, связанного с заказом.",
            "format": "int64"
          },
          "stock": {
            "description": "Информация о складе, связанном с заказом.",
            "allOf": [
              {
                "$ref": "#/components/schemas/StockDto"
              }
            ]
          },
          "orderItems": {
            "type": "array",
            "description": "Список товаров в заказе.",
            "items": {
              "$ref": "#/components/schemas/SellerOrderItemDto"
            }
          },
          "place": {
            "type": "string",
            "description": "Информация о месте обработки заказа."
          },
          "invoiceNumber": {
            "type": "integer",
            "description": "Номер накладной",
            "format": "int64"
          },
          "timeSlot": {
            "$ref": "#/components/schemas/TimeSlotDto",
            "description": "Интервал времени, выделенный для отгрузки заказа."
          },
          "dropOffPoint": {
            "$ref": "#/components/schemas/DropOffPointShortDto",
            "description": "Краткая информация о пункте сдачи заказа."
          },
          "scheme": {
            "type": "string",
            "description": "Тип заказа",
            "enum": [
              "FBS",
              "DBS",
              "FBO"
            ]
          },
          "deliveryInfo": {
            "$ref": "#/components/schemas/SellerOrderDeliveryInfoDto",
            "description": "Информация о доставке заказа."
          }
        }
      },
      "DropOffPointShortDto": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Идентификатор пункта приема",
            "format": "uuid",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "address": {
            "type": "string",
            "description": "Адрес пункта приема",
            "example": "г. Ташкент, Сергелийский район, массив Куйлюк, ул. Ханабад 2/1"
          },
          "type": {
            "type": "string",
            "description": "Тип пункта доставки",
            "example": "ISSUE_POINT",
            "enum": [
              "UNKNOWN",
              "STOCK",
              "ISSUE_POINT",
              "UZ_POST",
              "UCELL",
              "NOT_UZUM",
              "PHOTO_STUDIO",
              "CROSS_DOCK",
              "DROP_OFF_POINT",
              "SORT_CENTER"
            ]
          }
        },
        "description": "Пункт приема"
      },
...
      "GenericResponseSellerOrderDto": {
        "type": "object",
        "description": "Общий ответ, содержащий информацию о конкретном заказе продавца.",
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/SellerOrderDto"
          },
          "errors": {
            "type": "array",
            "description": "Список ошибок.",
            "items": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Метка времени ответа."
          },
          "trace": {
            "type": "string",
            "description": "Трассировочная информация для отладки."
          },
          "error": {
            "type": "string",
            "readOnly": true,
            "example": "null",
            "description": "Описание ошибки, если она произошла."
          }
        }
      },
...
      "GenericResponse": {
        "type": "object",
        "properties": {
          "payload": {
            "type": "object",
            "description": "Response payload"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "trace": {
            "type": "string"
          },
          "error": {
            "type": "string",
            "readOnly": true,
            "example": "null"
          }
        }
      }
    }
  }
}